#35: Rename tflayer.python package
