+++ tf-data-code/src/main/kotlin/edu/wpi/axon/tfdata/code/PythonUtil.kt
@@ -1,3 +1,3 @@
-package edu.wpi.axon.tflayer.python
+package edu.wpi.axon.tfdata.code
 
 fun boolToPythonString(bool: Boolean): String = if (bool) "True" else "False"
